updated the hostile movement.
Updated the Hostile movement. Monsters wont turn back anymore unless they walk into a dead end.

+++ Assets/Scripts/Entities.Impl/Player/PlayerController.cs
@@ -179,7 +179,7 @@ public class PlayerController : EntityAbs, PlayerIfc {
         }
         else {
             SetDirection(movableDirection);
-            StartCoroutine(Move(movableDirection, m_status.Speed));
+            StartCoroutine(Move(movableDirection, (int)m_status.Speed));
         }
     }
